plugin test at one place
